Expand infra-sk/themes.scss into a theme for all *.skia.org apps.
The default dark theme colors from the material design web site
are garish and everyone seems to have chose different colors as
they diverged from that theme.

This CL is the beginning of creating a common theme that can be
used on all of our sites.

It first defines buttons that actually look 3D and have an animation
when they are depressed. It also includes styles for buttons,
selects, details/summary, tables, and scrollbars.

Also:
 - Tightens up some CSS selectors to keep other apps unchanged.
 - Adds rules to app-sk so login-sk appears.

Change-Id: I1800a80e3332fcf264357ceeac8416c133dd1208
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/495420
Reviewed-by: Ravi Mistry <rmistry@google.com>
Commit-Queue: Joe Gregorio <jcgregorio@google.com>