Highlight hovered-over variables in the Debugger.
The code editor now has a mouse-move handler that tracks the word that
the mouse is currently hovered over. If the mouse is hovered over a
variable name that exists in the variable pane below, that line in the
variable pane is highlighted.

This doesn't actually attempt to parse the code, so it will trigger on
semantic mismatches, such as mousing over words in comments that contain
a variable's name, e.g.:

    myVar *= 2; // Double myVar.

In practice this isn't really a problem.

Change-Id: Ie89897c32b9d7fab3f479b0fdc357a2e73571f5a
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/496382
Reviewed-by: Joe Gregorio <jcgregorio@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>