[gold] fuzzy_test.go: Fix test cases that break under Go 1.18.
Starting Go 1.18, draw.Draw[1] includes some optimizations that, as a consequence, change the way in which transparent pixels are copied from the source to the destination image. If the alpha channel of a pixel is 0, the RGB channels will be set to 0, regardless of their original value. Before Go 1.18, the behavior was to copy the pixels as-is. This breaks fuzzy_test.go, which uses golden images with pixels where the alpha channel is zero and the RGB channels are non-zero.

Take for example the following Go snippet:

	a := image.NewNRGBA(image.Rect(0, 0, 1, 1))
	a.Set(0, 0, color.NRGBA{0xFF, 0xFF, 0xFF, 0})

	b := image.NewNRGBA(a.Bounds())
	draw.Draw(b, a.Bounds(), a, a.Bounds().Min, draw.Src)

	fmt.Printf("a: %v\n", a)
	fmt.Printf("b: %v\n", b)

Under Go 1.17.8, this prints out the following:

	a: &{[255 255 255 0] 4 (0,0)-(1,1)}
	b: &{[255 255 255 0] 4 (0,0)-(1,1)}

However, under Go 1.18 the output is different:

	a: &{[255 255 255 0] 4 (0,0)-(1,1)}
	b: &{[0 0 0 0] 4 (0,0)-(1,1)}

I made the above example available here: https://go.dev/play/p/IpAU8lZkqJU (uses the latest Go version, 1.18 at the time of writing).

This CL fixes the broken test cases by using non-zero alpha channels where appropriate.

[1] https://pkg.go.dev/image/draw@go1.18#Draw

Bug: skia:13063
Change-Id: I1330234611a9c9b9dda1c86b6ff12b2493849410
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/521336
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Leandro Lovisolo <lovisolo@google.com>