Fix data race surfaced after Go 1.18 upgrade.
The data race occurs because we simultaneously read the `closers` array on log_parser.go line 324, and write to it on line 286. The fix is to populate the array before launching the goroutine that reads from it.

I'm not sure what changed in Go 1.18 that makes the data race detector go off, but hopefully Infra-PerCommit-Race will go back to green after this CL lands.

Bug: skia:13063
Change-Id: Ie31632af5c633a97ee15e56d065e50f327fd60f8
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/523820
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Leandro Lovisolo <lovisolo@google.com>