[npm-audit-mirror] Add library that handles allow lists from the config
This reads the top-level packages specified in the config file and adds
all of their transitive dependencies to the allow list.

It will only find dependencies of packages that do not use semver, because we do not know the exact version to use for those packages when querying the global NPM registry.

Bug: skia:13057
Change-Id: I81952cf710838e7984354484f81efc17b0e96160
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/530566
Reviewed-by: Joe Gregorio <jcgregorio@google.com>