[npm-audit-mirror] Add ability to check for audit issues and automatically file bugs
This runs `npm audit` on a supported project's package.json file. If there are any high severity issues reported then it automatically files a monorail issue.
The monorail issue is only filed if one does not already exist, and if the last monorail issue for this project was closed more than a quarter ago.

Bug: skia:13057
Change-Id: I94fd459343f32b53454ddd26b7a8d9b738ae2089
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/530567
Reviewed-by: Joe Gregorio <jcgregorio@google.com>