[npm-audit-mirror] Add checks manager
The checks manager first validates that it is a package tarball request
and if it is then checks if the package is already installed, is a
trusted scope, is in the allowlist. If any none of those is true then it
performs security checks on the package.

Bug: skia:13057
Change-Id: I9344ce99ebad89014698a1326f7768d89daddaeb
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/530574
Reviewed-by: Joe Gregorio <jcgregorio@google.com>
Commit-Queue: Ravi Mistry <rmistry@google.com>