[skfe] Start setting up for managing our domain names via GCP.
The current DNS Zone management system is very slow to update,
so this starts the process of moving DNS record management
out of the current system into Cloud DNS.

This CL duplicates our current records in Cloud DNS, and adds
scripts and unit tests to confirm everything is configured
correctly.

Note that the Zone that is setup by this CL is not currently 'live'.
That is, we still point to the current system that manages
our Zone, and will have to make further changes to make
this go live. See also go/skia-dns.

Change-Id: I49ea6683c8c5d2caf781b3feacb386e1a854b51c
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/538043
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Joe Gregorio <jcgregorio@google.com>