[infra] Switch from python gsutil to golang gsutil
I do not think gsutil is used (otherwise we would be seeing
issues with Python 3.8+) in the infra CI, but it is on the
Skia one (e.g. uploading Gold images). Thus, I removed it
from the CI tasks here.

In order to make the switch on the Skia side (to unblock
only using Python3 on the CI), we need to stop using the
Python version (which was not-updated since 2020) and
use the more recent golang-based version (updated as
recently as last month).

https://chrome-infra-packages.appspot.com/p/infra/tools/luci/gsutil

This avoids https://github.com/GoogleCloudPlatform/gsutil/issues/961
which I saw when trying to land https://skia-review.googlesource.com/c/skia/+/538636

Change-Id: Ib9364d7ca9f7066838ed00b1f51c206b393eb46c
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/538818
Auto-Submit: Kevin Lubick <kjlubick@google.com>
Reviewed-by: Eric Boren <borenet@google.com>