[bazel] Remove skia-switchboard utils from RBE toolchain container image.
The skia-switchboard utilities weren't used by any of the non-manual tests. It's probably OK to remove them, and if we need them in the future, we can re-introduce them via a hermetic repository rule.

This CL also removes package binutils, which was either unused, or is now being pulled as a dependency of the clang-11 Debian package.

Bug: skia:13293
Change-Id: I7afcc3c96ab4a5af6d13b40c37d4846ee5b83829
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/548231
Reviewed-by: Joe Gregorio <jcgregorio@google.com>
Commit-Queue: Joe Gregorio <jcgregorio@google.com>
Auto-Submit: Leandro Lovisolo <lovisolo@google.com>