[prober] Fix building of allprobersk.json
allprobers.json wasn't being built since the destination
directory it was pointing to was inside the bazel build
directory.

Then fix how the tree is walked because now the file path
is absolute instead of relative.

Change the 'name' in skia_app_container so the
new image gets pushed correctly.

Finally remove allprobersk.json before rebuilding so
a failure to build the file will cause the build to fail.
As opposed to before this CL, where it would silently
fail and push the old allprobersk.json.

Change-Id: I802fa736f945cb52a0f4c575f9cb4c662a2ff89a
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/551436
Reviewed-by: Ravi Mistry <rmistry@google.com>
Commit-Queue: Joe Gregorio <jcgregorio@google.com>