Update github.com/jackc/pgx/v4 to v4.16.1.
We currently use a version from 2020, and the version
of pgconn that the current version relies upon has
many fixes for context timeouts.

I am currently seeing hangs in pgx far longer
than the context timeout should allow, but only
on the android-x instance, so my hope is that this
will fix the issue.

Bug: skia:13455
Change-Id: Iea5e6cbf16b773abb9b96aa17e1dc7043a71d3f9
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/552066
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Joe Gregorio <jcgregorio@google.com>