Add genpromcrd.
A command-line application for producing CRDs for Managed Prometheus.

$ go run ./promk/go/genpromcrd  --help
usage: genpromcrd --directory=[k8s-config checkout dir] [options]
options:
  -directory string
        The directory that contains a checkout of k8s-config.
  -dryrun
        If true then just print the names of the files that would be written.
  -logtostdout
        If true then write logging on stdout.

The genpromcrd cmd runs over all Deployments and StatefulSets and
writes out Managed Prometheus CRDs for both scraping and alerting.
For example, given the following file in the git repo that contains
all the cluster config:

        k8s-config/
        ├── monitoring
        │   └── appgroups
        │       └── perf.yml
        └── skia-infra-public
            └── perf.yml

All the Rules files for alerts to run for all Deployments and
StatefulSets are held under /monitoring/appgroups and the name
of the file before the '.yml' corresponds to an appgroup label.

Since perf.yaml resides inside a directory associated with a
cluster, the Deployment there runs in the namespace 'somenamespace',
and has .template.label.appgroup=perf, a new file will be written to:

   skia-infra-public/perf_somenamespace_appgroup_alerts.yml

which is a modified version of /monitoring/appgroups/perf.yaml, updated
to scrape the deployment in the correct namespace, and it will also
contain 'absent()' alerts for all the alerts defined in 'perf.yml'.

The list of directories processed are defined in:

    //kube/clusters/config.json



Bug: skia:13542
Change-Id: If4c1f6b281a236a3613ca4b65c6279d0be90e3d8
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/561420
Reviewed-by: Ravi Mistry <rmistry@google.com>
Commit-Queue: Joe Gregorio <jcgregorio@google.com>