[codesize] Display missing data from sections like .rodata
In order to do this, a bit more de-duplication of symbols needed to
be done, as some symbols can show up in multiple sections.

This also adds a maximum children to parent setting. Children smaller
than the top 200 will be folded together to keep the treemap from getting
too much to render or read.

This also fixes a small annoyance where not all third_party code would
be grouped together.

Bug: skia:12151
Change-Id: Ie9769e05da5e1ead1bd84acf53f1e8763876dec9
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/571409
Reviewed-by: Brian Osman <brianosman@google.com>