Display multiple commits per point on the explore page.
When a point is clicked on display all the commits that went into that data
point, excluding the most recent commit with data.

Note that this behavior is turned off for instances that run with
alerts.DefaultSparse == true, because in those instances the data is so sparse
the number of commits displayed could be huge (in the thousands).

Bug: skia:11286
Change-Id: Ib580c9b42122b217dbf7912f34cad03d3fda4a72
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/583747
Commit-Queue: Joe Gregorio <jcgregorio@google.com>
Reviewed-by: Ravi Mistry <rmistry@google.com>