[perf] Speed up queries by counting and restricting.
Doing the merge will be faster with a smaller number of
traces returned. They idea is to first do count(*) queries
for each key=[values] part of a query, and use the smallest
response to narrow down the number of trace ids to a small number.

Then issue the queries for all the rest of the key=[values], but now also include a restriction for the trace_ids.

SELECT
    count(*)
FROM
    Postings
WHERE
    tile_number = 0
    AND key_value = 'arch=x86';

In this example now that we've determined arch=x86 to return the smallest number of traces we retrieve all those trace_id's.

SELECT
    encode(trace_id, 'hex')
FROM
    Postings
WHERE
    tile_number = 0
    AND key_value = 'arch=x86';


Then use those trace_id's to constrain the other queries.
In practice don't bother with the IN clause if the number
of trace_id's returned from the count(*) are too large, >1,000 as a cutoff.

SELECT
    encode(trace_id, 'hex')
FROM
    Postings
WHERE
    tile_number = 0
    AND key_value = 'config=565'
    AND trace_id IN (
        '\xfe385b159ff55dca481069805e5ff050',
        '\x277262a9236d571883d47dab102070bc'
    );

Change-Id: I194b7f75b631fb60aa76e8d599e56e13faf8040c
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/589658
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Joe Gregorio <jcgregorio@google.com>