Remove cipd_git dep from git_common
After this CL, there is not an implicit dependency from git_common
to the version of git downloaded from CIPD because
git_common.FindGit() does not use that binary by default for
Bazel tests.

Instead, tests which need git now add a GitFinder to the context
they pass into git_common.FindGit (or any function which uses
that) which allows them to mock out using git or specify
the one from CIPD.

The default behavior is still to use git on the PATH, so as not
to affect production services. Tests which use this default
behavior will fail (because it is non-hermetic).

git_common.FindGit does not cache nor use the cached results
if the context has an override, to prevent different tests
from interfering with each other (learned the hard way).

I did most of this CL on my Mac, thus the addition of
exec_mac.go to make tests build on a Mac without using RBE.

Suggested review order:
 - go/git/git_common/
 - task_scheduler/go/tryjobs/ to see a small reorganization
   of test deps only in the test executable, not the whole
   package.
 - all other files to see lots of additions of
   git_common.WithGitFinder(context.Background(), cipd_git.FindGit)

Change-Id: Id7ac817b4b72b9f063a0427dff33b9479a247830
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/590555
Reviewed-by: Eric Boren <borenet@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>