[machine] Give TMM the ability to run foundry_bot in the background.
This is the first of three "Test Machine Monitor Changes Needed" items
from go/skolo-rbe. We may need to arrange for a service account key or
other auth to be passed in in a later commit. Args to Foundry Bot may
become customizable as the need arises, but for now they are hard-coded.

Also make a few facilitating improvements to executil:

* Make sure FakeCommandsReturned() doesn't let the command index change
  while we're reading it so we can have the main testing thread reliably
  observe the Foundry-Bot-launching goroutine.
* Add IsCallingFakeCommand() to boost concision and keep the details of
  how fake commands are implemented within executil.

Also revise a comment that refers to a non-existent util.WithCreate().

Change-Id: I8f8b5a19dc8fef79023f8efb7c4be323c2997357
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/595477
Commit-Queue: Erik Rose <erikrose@google.com>
Reviewed-by: Kevin Lubick <kjlubick@google.com>