Adds the sser Go module.
This module is for handling Server-Sent Events from a
ReplicaSet in kubernetes.

This CL adds the interfaces, mocks, and implementations
for handling such connections from the server side.

Additionally it adds an example application that is deployed
to prod as sserexample.skia.org to demonstrate that this
works with our current prod environment, including
interoperating well with auth-proxy.

Change-Id: I10e1a237ad0746f185c5c77ac954cb23de9c5511
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/596588
Reviewed-by: Erik Rose <erikrose@google.com>
Commit-Queue: Joe Gregorio <jcgregorio@google.com>