Add clipboard-sk element.
Displays a copy-content icon and when clicked copies the contents of the 'value' attribute into the user's clipboard. Also displays a tooltip letting the user know the value was copied.

If the value to be copied is expensive to calculate then compute the value in the `calculatedValue` function. See `clipboard-sk-demo.ts` for an example.

Bug: skia:13929
Change-Id: I2e385760f289d4184c642316152acdbcefef291e
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/607865
Auto-Submit: Joe Gregorio <jcgregorio@google.com>
Reviewed-by: Ravi Mistry <rmistry@google.com>
Commit-Queue: Joe Gregorio <jcgregorio@google.com>