[bazel] Remove dependency on depot_tools
This is basically the same as the Skia implementation [1], with
new support for post-install commands. At some point, we'll have
a single implementation in a common repo (same for our C++ toolchain).

This improves things by 1) removing a dependency on an external
binary; 2) improves hermeticity by having Bazel verify the files
we download (using the sha256 hash provided by CIPD); and 3)
having a mirror in case CIPD servers go down.

The mirror was uploaded using [2]; I downloaded the zip files
from the CIPD UI and using commands like:
    go run ./bazel/gcs_mirror/gcs_mirror.go --file ~/Downloads/git-windows-amd64.zip --sha256 9caaf...

As long as I was changing the rules, I updated the names of
these packages to include OS *and* architecture names, as we
have been doing elsewhere [3].

[1] https://github.com/google/skia/blob/0b87275d13a0dd6152f2214acbfefb9f5dcc367b/bazel/cipd_install.bzl
[2] https://github.com/google/skia/blob/0b87275d13a0dd6152f2214acbfefb9f5dcc367b/bazel/gcs_mirror/gcs_mirror.go
[3] https://github.com/google/skia/blob/0b87275d13a0dd6152f2214acbfefb9f5dcc367b/WORKSPACE.bazel#L263

Change-Id: Iee987ce2021fcf1039d418e124cec5a6502b8486
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/613743
Commit-Queue: Kevin Lubick <kjlubick@google.com>
Reviewed-by: Joe Gregorio <jcgregorio@google.com>