[skottie] Allow local servers to run on Macs
This requires updating puppeteer, which requires updating
typescript, which requires fixing several tests and type
annotations.

This uses Puppeteer v19.2.2 (latest that was in our npm mirror
at the time of writing). There was a major change in v19 about
where npm install downloads the chrome binary. It defaults to
a cache directory in the user's home directory, which is not visible
by Bazel, especially not RBE.

With some careful configuration (see //.puppeteerrc.js and
puppeteer-tests/BUILD.bazel) we can make that cached Chrome
available to Bazel for both local and remote testing. This
was a bit annoying to work out because kjlubick@ could not
figure out how to get any logs from the config file to
be visible (e.g. inspect environment variables). Maybe it
is possible for the config system used by Puppeteer [1].

[1] https://www.npmjs.com/package/cosmiconfig

Change-Id: I905e4f4d0a5e992644687e012ec2f8591e305180
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/611018
Commit-Queue: Kevin Lubick <kjlubick@google.com>
Reviewed-by: Joe Gregorio <jcgregorio@google.com>