[gold] JSON5 config: define groupings; use them in /json/v1/groupings RPC.
This CL aims at solving spurious test failures in the Skia and Chrome CIs due to goldctl reporting "grouping params for corpus X are unknown".

The /json/v1/groupings RPC reads groupings from Gold's status cache[1]. When the gold_frontend service starts serving, the status cache is initially empty. Then, a separate goroutine attempts to populate the status cache with the results of a SQL query[2]. It turns out that this query can be slow for large Gold instances (see "getCorporaStatuses" trace here[3]). I ran this query ad-hoc against the Skia and Chrome databases, and it took ~13 minutes and ~10 seconds, respectively.

My hypothesis is that the spurious test failures occur when goldctl issues a /json/v1/groupings RPC that is served by a Gold instance in the first moments after going online, during which the status cache might be empty. If so, the /json/v1/groupings RPC returns an empty dictionary of param keys by corpus, which causes goldctl to fail[4].

My proposed fix is to declare the groupings in advance in an instance's JSON5 config, and to return those if the status cache is empty. For now this mechanism is optional. This CL adds grouping definitions to the JSON5 configs of the Skia and Skia Infra instances to begin with. If we like this approach, I'll repeat this for all the other instances.

A benefit of this approach is that we could leverage the JSON5 config in the future to add support for groupings other than (name, source_type).

Another potential benefit is that we could use this to return a non-empty (but not fully populated) /json/v2/trstatus RPC[5] response while the status cache is initially empty, which is part of the reason why the Skia Gold's home page errors out during the first few minutes after redeploying gold_frontend.

Recommended review order:

 - //golden/cmd/gold_frontend/gold_frontend.go
 - //golden/go/web/web.go
 - //golden/go/web/web_test.go
 - //golden/k8s-instances/skia/skia-frontend.json5
 - //golden/k8s-instances/skia-infra/skia-infra-frontend.json5
 - //golden/gold_launcher.bzl
 - //golden/BUILD.bazel

[1] https://skia.googlesource.com/buildbot/+/f35cc15b2bdbe70016c5c064a85abe3202ba191e/golden/go/web/web.go#116
[2] https://skia.googlesource.com/buildbot/+/f35cc15b2bdbe70016c5c064a85abe3202ba191e/golden/go/search/search.go#3968
[3] https://pantheon.corp.google.com/traces/list?project=skia-public&pageState=(%22traceFilter%22:(%22chips%22:%22%255B%257B_22k_22_3A_22SpanName_22_2C_22t_22_3A10_2C_22v_22_3A_22_5C_22getCorporaStatuses_5C_22_22%257D%255D%22))&tid=00261fff1c735cf65b56be5e8fa6cdbf
[4] https://skia.googlesource.com/buildbot/+/f35cc15b2bdbe70016c5c064a85abe3202ba191e/gold-client/go/goldclient/goldclient.go#418
[5] https://skia.googlesource.com/buildbot/+/f35cc15b2bdbe70016c5c064a85abe3202ba191e/golden/go/web/web.go#1338

Bug: chromium:1403684
Change-Id: I400bee4683d3a42ff4cb13abea2bf62f48a14c1e
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/632719
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Leandro Lovisolo <lovisolo@google.com>