Fix /archive/ issues the best we can, file bugs for others
On Jan 31, 2023, GitHub changed their compression algorithm
(and eventually rolled it back) for the auto-generated source
zip/tar.gz files. This broke any Bazel builds which used them
because the sha256 hashes changed.

An ideal fix would be to use a source archive that has been manually
added to the GitHub release (with our mirror as a backup). Those
user-provided archive files will *not* change in the future.

However, not every one of our deps has such releases. I've filed
bugs with those which do not and in the mean-time changed the
mirror logic to use our mirror first (which we know will not
change) so when GitHub rolls this out in the future (probably
not for a while), we won't break.

Change-Id: I102926ca39a07f435cfd46f266dba591ce1d9c40
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/635613
Reviewed-by: Joe Gregorio <jcgregorio@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>