[gce] swarming_vm.go: Stop writing Windows setup/startup scripts to disk.
Currently, setting up a Windows VM involves reading the setup and startup script templates from disk, replacing a placeholder with the plaintext chrome-bot password, then writing these scripts to disk. Then, the scripts are read from disk and their contents are used to populate a GCE API call to create the VM.

This CL skips writing those files to disk by:
 - Making the gce.Instance struct store the script contents, rather than a path to the script.
 - Populating said struct directly with the results of replacing those placeholders in the Windows setup and startup scripts.

I tested this CL by creating a Windows and a Linux VM with swarming_vm.go, and a "worker" machine with ct_vm.go. The latter failed due to an issue with the setup script itself, and not the mechanism by which ct_vm.go delivers setup scripts to VMs: http://go/paste/4688995635167232. The fact that the setup script made it into the VM tells me that this CL did not break ct_vm.go, so it's probably safe to land.

Bug: skia:14074
Change-Id: I497181fd689a1534bff63835856cacbc7c90baa2
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/637660
Reviewed-by: Eric Boren <borenet@google.com>
Commit-Queue: Leandro Lovisolo <lovisolo@google.com>
Reviewed-by: Ravi Mistry <rmistry@google.com>