[scrap] add utility program to set default SkSL scrap ImageURL
The stock SkSL images, which are referenced by most shader
scraps, need to be moved from /dist/ to /img/. This is
to allow them to be fetched cross-origin. This utility
program modifies the @default named SkSL scrap, which is
used by the shaders web app as a template for newly created
shader scraps.

A follow-on change will convert all scraps from /dist/ to
/img/.

Bug: skia:14091
Change-Id: Ifaa6a7222685f41d5a49d028dde0c368c32a2592
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/639336
Reviewed-by: Joe Gregorio <jcgregorio@google.com>
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Chris Mumford <cmumford@google.com>