[machines] Implement GetFreeMachines in store.Store.
This also updates the schema for the Description table
to include a computed column, `runningTask`, which is true
if a task is running. This is needed because you can't create
a plain INDEX on a JSONB type column, so we can't quickly
select rows with a NULL task_request.

Also updates the schema migration test to accommodate multiple
tables.

Bug: b/265796964
Change-Id: I9608f3eb7372906a108b2ea957d3436445a21144
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/640156
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>
Commit-Queue: Joe Gregorio <jcgregorio@google.com>