Create DOCKER make variable
Create a $(DOCKER) make variable which explicitly uses the path returned
by `which docker`, or "docker" if empty. This works around an issue
where docker is installed on a system, but when run via make, will
fail with the following error:

    docker pull gcr.io/skia-public/skia-release:prod
    make: docker: Permission denied
    make: *** [Makefile:27: build-fiddler] Error 127

This failure is triggered when running make after sourcing
//skia/third_party/externals/emsdk/emsdk_env.sh. That script
modifies the PATH which causes make to try executing
//skia/third_party/externals/emsdk/docker (which is a directory)
instead of the installed docker executable (/bin/docker on Linux).

Change-Id: I9e1716769e48de8acc3bc166802cae1ac686a564
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/649544
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Chris Mumford <cmumford@google.com>
Reviewed-by: Ravi Mistry <rmistry@google.com>