Change the keyRegex and paramRegex to only check the key format, avoid ',' or '=',  and avoid ';' to avoid SQL injection attack.
Since the ingestor has already applied the force valid based on the invalidParamCharRegex, the following DB read/write or other operations should always deal with valid parameter sets. In this case, the key (traceName) should always be valid as well.
So, this CL is to change the keyRegex and paramRegex to only check the key format, avoid ',' or '=' and avoid ';' to avoid SQL injection attack

Change-Id: Ib472f9f269ecfed576ee82f1275e771254dfe31b
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/700755
Commit-Queue: Alex Sun <sunpeng@google.com>
Reviewed-by: Joe Gregorio <jcgregorio@google.com>