[cabe] add inferred_experiment_spec to GetAnalysisResponse proto
If the client asks cabe to analyze a pinpoint job but does not include
an experiment spec to tell cabe exactly what to look for, cabe will
attempt to infer what that spec should be, by applying some heuristics
to swarming task metadata and contents of benchmark task output files.

This change adds that inferred experiment spec as a new field on the
GetAnalysisResponse proto that cabe returns, if there was no such
experiments_spec provided by the client in the original
GetAnalysisRequest proto. If the client provided a spec in the
GetAnalysisRequest, cabe leaves the inferred_experiment_spec field
empty.

Change-Id: I8ece2b35fb62d48e9816df3fbc5300e35fd19ccc
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/705577
Reviewed-by: Yuan Huang <yuanhuang@google.com>
Commit-Queue: Sean McCullough <seanmccullough@google.com>