[perf] Better tracing that includes the time to write traces.
The entire point of this CL is to add these lines:

   ctx, span := trace.StartSpan(ctx, "ingest.parser.processSingleFile")
   defer span.End()

But to get that to work the insides of the for loop in `worker`
needed to be broken out into its own function, processSingleFile,
and that required factoring out all the variables that processSingleFile needs into a struct.

This does make `worker` much shorter, and makes all
the variables to processSingleFile explicit.

The new context at the beginning of processSingleFile is there
so the traces for reading the file from GCS, and the traces
for writing data into CockroachDB show up as child traces
of processSingleFile.

Change-Id: Id907660e55e8703cd703b7f6b77aef591b87a0b2
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/706521
Commit-Queue: Joe Gregorio <jcgregorio@google.com>
Reviewed-by: Alex Sun <sunpeng@google.com>