Move Gold to auth-proxy.
Removes all login flows from the Gold executable, replaces it
with proxylogin that gives access to the logged in user and their
roles.

Also updates the Gold frontend template to start running
auth-proxy in front of each gold instance.

Note that since auth-proxy is just enforcing the same authentication
that the Gold frontend instance is also enforcing this change
is fine even if the version of Gold isn't updated.

Here is how the changes to the template will be rendered:

  https://skia-review.googlesource.com/c/k8s-config/+/708779

Change-Id: Ic245e919b23eec5f040a7c2e4cd86a6470bb6838
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/707718
Commit-Queue: Joe Gregorio <jcgregorio@google.com>
Reviewed-by: Kevin Lubick <kjlubick@google.com>