[build-images] Add tests for updateRefs
While running the tests on my Mac, I realized the git auth
step was really generating a token based on the
GOOGLE_APPLICATION_CREDENTIALS environment variable. This is not
set on said Mac and thus caused failures. Thus, I made that
swappable via the passed in context and made all tests in build-images
use it.

I also needed to make the creation of a random ChangeId for the
rubberstamper interceptable, to make tests deterministic.

These basic tests for update-refs don't cover the scenario when
a stage manager is set, if pub sub is configured, nor when using
gitiles to construct the commit message. Those can be added in
future CLs, as they will require more clients to be injected
either via function params or on context.Context (former is preferred).

Bug: b/286377352
Change-Id: If1f89a6c90b82da61554c219cdca8d12a1f09e51
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/725502
Commit-Queue: Kevin Lubick <kjlubick@google.com>
Reviewed-by: Eric Boren <borenet@google.com>