Minor fixes required to re-deploy Gold
Flutter-engine's periodic tasks had gotten messed up in that
the CL tracker was trying to compare against a non-existent commit.

I noticed the CL mirror should be using the "faster" mirror,
not the chrome one, so hopefully that will help avoid issues
where GitHub and Gitiles got out of sync. I'm leaving in
the debugging URL to ease future issues with gitiles.

Golang recently started panicking if time.NewTicker was called
with duration 0, so this prevents that as well.

Change-Id: If00b90d86b15faac37065dccbb604e963eba14e7
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/753549
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>