[gold] Fix bug preventing Gold from diffing digests uploaded from CLs.
In Impl.getExpectationsForDigest(), we query the Expectations and SecondaryBranchExpectations tables for expectations for the given digest. This query uses the COALESCE[1] function to combine them, such that the CL expectation takes precedence over the primary branch expectation, defaulting to untriaged ('u') if none is found.

There is a bug in this query, which happens when no expectations are found on either table. Because the set of rows returned by the ExpectationsFromCL and ExpectationsFromPrimary common table expressions are both empty, the "SELECT COALESCE(...) FROM ExpectationsFromCL FULL OUTER JOIN ExpectationsFromPrimary ON ..." query returns 0 rows. This causes Impl.getExpectationsForDigest() to incorrectly error out.

The fix is to ensure we return a NULL value if both expectations are absent, rather than no value at all, and feed that to COALESCE.

On patchset 1, I added a test that reproduces the bug.

Patchsets 2+ fix the bug and update a test that broke because of this change.

[1] https://www.cockroachlabs.com/docs/v23.1/scalar-expressions#coalesce-and-ifnull-expressions

Bug: b/297010848
Change-Id: I4ad3f0b1584dc5c84e27e76e7a441c142a99b9ec
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/753414
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Leandro Lovisolo <lovisolo@google.com>