[cabe] handle “unequal number of tasks for treatment and control” better
A motivating example for this fix:
https://pinpoint-dot-chromeperf.appspot.com/job/153eb621d60000

Only 7 of the 36 replicas produced the requested metric in both arms.
Many pairs had it in one arm but not the other, and many were didn't
include it in either arm.  This CL lets cabe's analyzer process the
results anyways, using whatever pairs did manage to produce the
metric from both tasks.

Running this command:
```
bazelisk run //cabe/go/cmd/cabe -- analyze --pinpoint-job 153eb621d60000 -benchmark rendering.desktop -workload Graphics.Smoothness.PercentDroppedFrames3.AllSequences
```
with this CL now produces this output instead of failing with an error:
https://paste.googleplex.com/6308328048164864


This change makes a few modifications to the analyzer logic so that it
can handle pinpoint tryjobs where one or both tasks run to completion
and produce valide perf_results.json files but nevertheless do not
include the requested benchmark/workload in the output file.

It makes no assumptions and does not try to verify that some minimum
number or fraction of task pairs are usable.  If there's more than
zero, it'll probably try to analyze it. We can circle back after this
lands and try to put some guardrails up, if we can establish some
heuristics about the minimums required.

There was one undocumented bug I discovered while fixing this one:
`armTask` pairing order gets screwed up if you have to toss out tasks
or task pairs.  So I fixed that issue, but it did affect some of
the calculations that are compared to hard-coded values in unit
tests. I updated the expected results to match what the code now
produces.

These differences affect the CI (and sometimes P-value) results but
not the medians for X and Y (aka the two experiment arms).  This
is consistent with keeping the same reported measurements for
each arm, but just running the hypothesis test using a different
(well, now it's *correct*) pairing order.  Same sets of values
result in the same medians, but the order change can affect the
CIs and P-values slightly.

Bug: chromium:1481203
Change-Id: Id5f670febb450d9269cec36dbc130449838923a3
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/759316
Reviewed-by: Jeff Yoon <jeffyoon@google.com>